requirements/base.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
click==6.6
coreapi==1.32.0
decorator==4.0.10
Django==1.9.8
django-dotenv==1.4.1
django-extensions==1.6.7
djangorestframework==3.4.0
dotenv==0.0.5
ipython==5.0.0
ipython-genutils==0.1.0
itypes==1.1.0
ordereddict==1.1
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
poster==0.8.1
pprintpp==0.2.3
prompt-toolkit==1.0.3
psycopg2==2.6.2
ptyprocess==0.5.1
python-dotenv==0.5.1
requests==2.10.0
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.2.2
Unirest==1.1.7
uritemplate==0.6
wcwidth==0.1.7
